+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Add new default `"auto"` setting for `RustExtension.py_limited_api`. [#137](https://github.com/PyO3/setuptools-rust/pull/137)
 - Support very verbose cargo build.rs output. [#140](https://github.com/PyO3/setuptools-rust/pull/140)
 
+### Changed
+- Switch to `tomli` dependency. [#174](https://github.com/PyO3/setuptools-rust/pull/174)
+
 ### Removed
 - Remove `test_rust` command. (`python setup.py test` is deprecated.) [#129](https://github.com/PyO3/setuptools-rust/pull/129)
 - Remove `check_rust` command. [#131](https://github.com/PyO3/setuptools-rust/pull/131)